requirements.txt
fastapi>=0.110
uvicorn>=0.27
langgraph==1.0.7
langchain>=0.2.0
langchain-google-genai>=0.1.0
google-generativeai>=0.5.0
python-dotenv
requests
pydantic>=2.0
